+++ src/Engine/Graphics/BSPModel.h
@@ -4,8 +4,8 @@
 #include <vector>
 #include <string>
 
-#include "Utility/Geometry/Plane.h"
-#include "Utility/Geometry/BBox.h"
+#include "Library/Geometry/Plane.h"
+#include "Library/Geometry/BBox.h"
 
 #include "FaceEnums.h"
