+++ src/Engine/Graphics/RenderEntities.h
@@ -6,8 +6,7 @@
 #include "Engine/Pid.h"
 
 #include "Library/Color/Color.h"
-
-#include "Utility/Geometry/Vec.h"
+#include "Library/Geometry/Vec.h"
 
 class Sprite;
 class SpriteFrame;
