+++ src/Engine/Graphics/Vis.h
@@ -1,11 +1,12 @@
 #pragma once
 
-#include "Utility/Flags.h"
-#include "Utility/Geometry/Plane.h"
-
 #include "Engine/Graphics/RenderEntities.h"
 #include "Engine/Objects/ActorEnums.h"
 #include "Engine/Pid.h"
+
+#include "Library/Geometry/Plane.h"
+
+#include "Utility/Flags.h"
 
 class BSPModel;
 struct ODMFace;
