+++ src/Engine/Graphics/Weather.h
@@ -2,7 +2,7 @@
 
 #include <array>
 
-#include "Utility/Geometry/Point.h"
+#include "Library/Geometry/Point.h"
 
 class Weather {
  public:
