+++ src/Engine/Snapshots/EntitySnapshots.h
@@ -2,9 +2,9 @@
 
 #include <array>
 
-#include "Utility/Geometry/Vec.h"
-#include "Utility/Geometry/Plane.h"
-#include "Utility/Geometry/BBox.h" // TODO(captainurist): Don't depend on BBox binary layout.
+#include "Library/Geometry/Vec.h"
+#include "Library/Geometry/Plane.h"
+#include "Library/Geometry/BBox.h" // TODO(captainurist): Don't depend on BBox binary layout.
 
 #include "Library/Binary/BinarySerialization.h"
