+++ src/Engine/SpawnPoint.h
@@ -4,7 +4,7 @@
 #include "Engine/Objects/ItemEnums.h"
 #include "Engine/Pid.h"
 
-#include "Utility/Geometry/Vec.h"
+#include "Library/Geometry/Vec.h"
 
 struct SpawnPoint {
     Vec3i vPosition;
