+++ src/Engine/TeleportPoint.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "Utility/Geometry/Vec.h"
+#include "Library/Geometry/Vec.h"
 
 /**
  * Describes target teleportation point of the party.
