+++ src/Library/CMakeLists.txt
@@ -6,6 +6,7 @@ add_subdirectory(Cli)
 add_subdirectory(Color)
 add_subdirectory(Compression)
 add_subdirectory(Config)
+add_subdirectory(Geometry)
 add_subdirectory(Image)
 add_subdirectory(Json)
 add_subdirectory(Lod)
