+++ src/Library/Image/CMakeLists.txt
@@ -11,5 +11,5 @@ set(LIBRARY_IMAGE_HEADERS
         PCX.h)
 
 add_library(library_image STATIC ${LIBRARY_IMAGE_SOURCES} ${LIBRARY_IMAGE_HEADERS})
-target_link_libraries(library_image PUBLIC library_color utility)
+target_link_libraries(library_image PUBLIC library_color library_geometry utility)
 target_check_style(library_image)
