+++ src/Media/Media.h
@@ -3,7 +3,6 @@
 #include <memory>
 #include <string>
 
-#include "Utility/Geometry/Rect.h"
 #include "Utility/Memory/Blob.h"
 
 class IAudioDataSource {
